+++ src/main/java/com/banchan/repository/QuestionDetailsRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
-public interface QuestionDetailsRepository extends JpaRepository<QuestionDetails, Integer> {
+public interface QuestionDetailsRepository extends JpaRepository<QuestionDetails, Long> {
 
-    CompletableFuture<List<QuestionDetails>> findALLByQuestionIdInOrderByQuestionIdAsc(List<Integer> questionIds);
+    CompletableFuture<List<QuestionDetails>> findALLByQuestionIdInOrderByQuestionIdAsc(List<Long> questionIds);
 }
